requirements.txt
Django
django-rest-framework
djangorestframework-jwt
django-allauth
django-rest-auth[with_social]